+++ src/app/modules/project/details/details.component.ts
@@ -16,22 +16,25 @@
  */
 import { environment } from 'src/environments/environment';
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Project } from 'src/app/models/domain/project';
 import { ProjectService } from 'src/app/services/project.service';
 import { AuthService } from 'src/app/services/auth.service';
 import { HighlightService } from 'src/app/services/highlight.service';
 import { HighlightAdd } from 'src/app/models/resources/highlight-add';
-import { BsModalService } from 'ngx-bootstrap/modal';
-import { ModalHighlightComponent, HighlightFormResult } from 'src/app/components/modals/modal-highlight/modal-highlight.component';
+import { BsModalService, ModalOptions } from 'ngx-bootstrap/modal';
+import { ModalHighlightComponent, HighlightFormResult } from 'src/app/modules/project/modal-highlight/modal-highlight.component';
 import { AlertConfig } from 'src/app/models/internal/alert-config';
 import { AlertType } from 'src/app/models/internal/alert-type';
 import { AlertService } from 'src/app/services/alert.service';
 import { switchMap } from 'rxjs/operators';
 import { User } from 'src/app/models/domain/user';
+import { Observable, EMPTY } from 'rxjs';
 import { HighlightByProjectIdService } from 'src/app/services/highlightid.service';
-import { ModalDeleteComponent } from 'src/app/components/modals/modal-delete/modal-delete.component';
+import { ModalHighlightDeleteComponent } from 'src/app/modules/project/modal-highlight-delete/modal-highlight-delete.component';
 import { Highlight } from 'src/app/models/domain/hightlight';
+import { ModalDeleteGenericComponent } from 'src/app/components/modals/modal-delete-generic/modal-delete-generic.component';
+
 
 /**
  * Overview of a single project
@@ -59,7 +62,8 @@ export class DetailsComponent implements OnInit {
     private highlightService: HighlightService,
     private modalService: BsModalService,
     private alertService: AlertService,
-    private highlightByProjectIdService: HighlightByProjectIdService
+    private highlightByProjectIdService: HighlightByProjectIdService,
+    private router: Router
   ) { }
 
   ngOnInit(): void {
@@ -132,7 +136,7 @@ export class DetailsComponent implements OnInit {
           return this.highlightService.post(highlightAddResource);
         })
       )
-      .subscribe((highlightFormResult: HighlightFormResult) => {
+      .subscribe(() => {
         const alertConfig: AlertConfig = {
           type: AlertType.success,
           mainMessage: 'Project was successfully highlighted',
@@ -166,7 +170,7 @@ export class DetailsComponent implements OnInit {
           }
         });
         const initialState = { highlights: results };
-        this.modalService.show(ModalDeleteComponent, { initialState });
+        this.modalService.show(ModalHighlightDeleteComponent, { initialState });
       }
     );
   }
@@ -180,14 +184,40 @@ export class DetailsComponent implements OnInit {
   }
 
   /**
-   * Method to display the edit project button based on the current user and the project user.
-   * @param project The project to check if the current user is the owner.
+   * Method which triggers when the delete project button is clicked.
+   * Displays the remove modal.
+   * Removes the project if modal returned true to confirm the delete.
    */
-  private determineDisplayEditProjectButton(): void {
-    if (this.currentUser == null || this.project == null || this.project.user == null) {
-      this.displayEditButton = false;
-    }
-    this.displayEditButton = this.project.user.id === this.currentUser.id;
+  public onClickRemoveProject(): void {
+    const modalOptions: ModalOptions = {
+      initialState: {
+        titleText: 'Delete project',
+        mainText: `Are you sure you want to delete the project, ${this.project.name}?`,
+      }
+    };
+    // Display modal
+    const modalRef = this.modalService.show(ModalDeleteGenericComponent, modalOptions);
+    // Map observable back to original type
+    const modalRefRemove = modalRef.content.remove as Observable<boolean>;
+
+    // Subscribe to remove event.
+    // Call the project remove service if true was returned.
+    modalRefRemove.pipe(
+      switchMap(deleteProject => {
+        if (deleteProject) {
+          return this.projectService.delete(this.project.id);
+        }
+        return EMPTY;
+      })
+    ).subscribe(() => {
+      this.alertService.pushAlert({
+        mainMessage: 'Removal of project was successful',
+        timeout: this.alertService.defaultTimeout,
+        dismissible: true,
+        type: AlertType.success
+      });
+      this.router.navigate(['project/overview']);
+    });
   }
 
   private formatTimestamps(highlightTimestamp: string): string {